+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 * nostr: add `Coordinate::has_identifier` ([Yuki Kishimoto])
 * pool: add `RelayPoolNotification::Authenticated` variant ([Yuki Kishimoto])
 * pool: add `RelayPool::save_subscription` ([Yuki Kishimoto])
+* sqlite/rocksdb/indexeddb: allow to open database with limited capacity ([Yuki Kishimoto])
 * sdk: add `Client::gift_wrap_to` and `Client::send_private_msg_to` ([reyamir])
 * sdk: add option to autoconnect relay on `Client::add_relay` method call ([Yuki Kishimoto])
 * sdk: add support to embedded tor client ([Yuki Kishimoto])